moved debug to component
