+++ main.py
@@ -1,13 +1,12 @@
 from datetime import datetime
-import json
-import pathlib
-import re
 import sys
 import matplotlib
 matplotlib.use('Qt5Agg')
 from PyQt5 import QtWidgets, QtGui, QtCore
 from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg as FigureCanvas
 from matplotlib.figure import Figure
+
+from qt_components.debug_button import DebugButtons
 
 from tinydb import TinyDB
 from packages.helper  import Helper
@@ -390,82 +389,20 @@ class HandicapUI(QtWidgets.QWidget):
         self.spiele_tabelle.setHorizontalHeaderLabels(["Datum", "Kurs", "Schläge"])
         layout.addWidget(self.spiele_tabelle)
 
-        #UNCOMMENT AFTER DEBUG
-        #self.setLayout(layout)
+
+        self.debug_buttons = DebugButtons(help, self)
+        layout.addWidget(self.debug_buttons)
+
+        self.setLayout(layout)
 
         self.neues_spiel_button.clicked.connect(self.oeffne_neues_spiel_dialog)
         self.neuer_kurs_button.clicked.connect(self.neuer_kurs_hinzugefuegt)
         self.kurs_löschen_button.clicked.connect(self.kurs_loeschen_dialog)
         self.export_scorecard_button.clicked.connect(self.oeffne_export_scorecard_dialog) #------------------------------------
 
-        self.update()
-        #DEBUG CODE DELETE BEFORE SHIPPING
-##################################################################################
-        self.lade_games_button = QtWidgets.QPushButton("Import Lade Games", self)
-        self.lade_courses_button = QtWidgets.QPushButton("Import Lade Courses", self)
-        self.lade_games_button.clicked.connect(self.add_lade_games)
-        self.lade_courses_button.clicked.connect(self.add_lade_courses)
 
-        self.truncate_button = QtWidgets.QPushButton("Clear DB", self)
-        self.truncate_button.clicked.connect(self.truncate_db)
-
-        layout.addWidget(self.lade_courses_button)
-        layout.addWidget(self.lade_games_button)
-        layout.addWidget(self.truncate_button)
-
-        self.setLayout(layout)
-
-        self.update()
-
-
-    def truncate_db(self):
-        help.games.truncate()
-        help.courses.truncate()
-        help.hcLog.truncate()
-        self.update()
-
-    def add_lade_games(self):
-        files = list(pathlib.Path("./test/real_data/games/").glob("Runde*.json"))
-        files.sort(key=lambda f: int(re.search(r'\d+', f.stem).group()))
-        files = list(pathlib.Path("./test/real_data/games/").glob("Runde*.json"))
-        files.sort(key=lambda f: int(re.search(r'\d+', f.stem).group()))
-        games = list()
-        for file in files:
-            with file.open() as file:
-                data = json.load(file)
-            games.append(data)
         
-        for game in games:
-            courseID = game["courseID"]
-            shots = game["shots"]
-            nineHole = game["is9Hole"]
-            pcc = game["pcc"]
-            cba = game["cba"]
-            gameDate = game["date"]
-            help.addGame(courseID, shots, nineHole, pcc, cba, gameDate)
         self.update()
-
-    def add_lade_courses(self):
-        files = list(pathlib.Path("./test/real_data/courses/").glob("Kurs*.json"))
-        files = list(pathlib.Path("./test/real_data/courses/").glob("Kurs*.json"))
-        courses = list()
-        for file in files:
-            with file.open() as file:
-                data = json.load(file)
-            courses.append(data)
-        
-        for course in courses:
-            help.addCourse(
-                course["courseID"],
-                course["course_rating"],
-                course["slope_rating"],
-                course["par"],
-                course["handicap_stroke_index"],
-            )
-#################################################################################
-
-
-
 
     def oeffne_neues_spiel_dialog(self):
         kurse = help.getAllCourseIDs()
